Move exported functions used in capabilityHelper.h from extern "C" to api.
