+++ capnp/__init__.py
@@ -52,6 +52,8 @@ from .lib.capnp import (
     _write_message_to_fd,
     _write_packed_message_to_fd,
     _Promise as Promise,
+    _init_capnp_api,
 )
 
+_init_capnp_api()
 add_import_hook() # enable import hook by default
